# API/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Grocery App",
    "description": "Grocery App with Flutter and Node.js",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/product": {
      "get": {
        "tags": ["Product"],
        "summary": "Get all products",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "required": false,
            "description": "Product Name",
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "Category ID",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page Number",
            "type": "integer",
            "default": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Page Size",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductServiceResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Product"],
        "summary": "Create Product",
        "parameters": [
          {
            "name": "productName",
            "in": "formData",
            "description": "Product Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "productShortDescription",
            "in": "formData",
            "description": "Product Short Description",
            "required": true,
            "type": "string"
          },
          {
            "name": "productDescription",
            "in": "formData",
            "description": "Product Description",
            "type": "string"
          },
          {
            "name": "productPrice",
            "in": "formData",
            "description": "Product Price",
            "required": true,
            "type": "number"
          },
          {
            "name": "productSalePrice",
            "in": "formData",
            "description": "Product Sale Price",
            "type": "number",
            "default": 0
          },
          {
            "name": "productSKU",
            "in": "formData",
            "description": "Product SKU",
            "type": "string"
          },
          {
            "name": "productType",
            "in": "formData",
            "description": "Product Type",
            "required": true,
            "type": "string",
            "default": "Simple"
          },
          {
            "name": "stockStatus",
            "in": "formData",
            "description": "Stock Status",
            "type": "string",
            "default": "In Stock"
          },
          {
            "name": "productImage",
            "in": "formData",
            "description": "Product Image",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductServiceResponse"
            }
          }
        }
      }
    },
    "/api/product/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Product Id",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Product"],
        "summary": "Get product by ID",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductServiceResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["Product"],
        "summary": "Update Product",
        "parameters": [
          {
            "name": "productName",
            "in": "formData",
            "description": "Product Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "Category",
            "required": true,
            "type": "string"
          },
          {
            "name": "productShortDescription",
            "in": "formData",
            "description": "Product Short Description",
            "required": true,
            "type": "string"
          },
          {
            "name": "productDescription",
            "in": "formData",
            "description": "Product Description",
            "type": "string"
          },
          {
            "name": "productPrice",
            "in": "formData",
            "description": "Product Price",
            "required": true,
            "type": "number"
          },
          {
            "name": "productSalePrice",
            "in": "formData",
            "description": "Product Sale Price",
            "type": "number",
            "default": 0
          },
          {
            "name": "productSKU",
            "in": "formData",
            "description": "Product SKU",
            "type": "string"
          },
          {
            "name": "productType",
            "in": "formData",
            "description": "Product Type",
            "required": true,
            "type": "string",
            "default": "Simple"
          },
          {
            "name": "stockStatus",
            "in": "formData",
            "description": "Stock Status",
            "type": "string",
            "default": "In Stock"
          },
          {
            "name": "productImage",
            "in": "formData",
            "description": "Product Image",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductServiceResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["Product"],
        "summary": "Delete Product",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductServiceResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Product": {
      "required": ["productName", "category", "productShortDescription", "productPrice", "productType"],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "productName": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "productShortDescription": {
          "type": "string"
        },
        "productDescription": {
          "type": "string"
        },
        "productPrice": {
          "type": "number"
        },
        "productSalePrice": {
          "type": "number",
          "default": 0
        },
        "productImage": {
          "type": "string"
        },
        "productSKU": {
          "type": "string"
        },
        "productType": {
          "type": "string",
          "default": "Simple"
        },
        "stockStatus": {
          "type": "string",
          "default": "In Stock"
        }
      }
    },
...
    "ProductServiceResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    }
  }
}